pyproject.toml
[project]
name = "leadsense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "httpx>=0.28.1",
    "mcp-server-fetch>=2025.1.17",
    "mcp[cli]>=1.5.0",
    "openai>=1.68.2",
    "openai-agents>=0.0.15",
    "playwright>=1.51.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
